core: defined 32-bit header load in the KL4E/KL3E decoder
The four header bytes were shifted as promoted ints, so a stream whose
first length byte was 0x80 or above shifted a value past INT_MAX:
undefined behaviour, and a hard stop under the UBSan build of
fuzz-parsers (-fno-sanitize-recover).  The bytes are cast to u32 first.

+++ src/core/nk_psp_kle.c
@@ -196,7 +196,8 @@ int decompress_kle(u8 *outBuf, int outSize, u8 *inBuf, int inSize, void **end, i
     in.cur = inBuf;
     in.end = inBuf + inSize;
     in.overrun = 0;
-    inputVal = (inBuf[1] << 24) | (inBuf[2] << 16) | (inBuf[3] << 8) | inBuf[4];
+    inputVal = ((u32)inBuf[1] << 24) | ((u32)inBuf[2] << 16) |
+               ((u32)inBuf[3] << 8) | (u32)inBuf[4];
     // Handle the direct copy case (if the file is actually not compressed).
     if (inBuf[0] & 0x80) {
         u8 *dataEnd = outBuf + inputVal;
